Add option -a to ammend commit
If the current HEAD has not been pushed and the -a option is given,
the new commit will be ammended to the old. This is designed to reduce
the proliferation of automatic commits for frequently-updated
repositories.

Add the option --all to git add to suppress a warning about future
changes to git:
warning: You ran 'git add' with neither '-A (--all)' or '--ignore-removal',
whose behaviour will change in Git 2.0 with respect to paths you removed.